+++ src/Debug.hs
@@ -9,27 +9,24 @@ module Debug
        , trace
        , traceM
        , traceId
-       , traceIO
        , traceShow
        , traceShowId
        , traceShowM
-       , notImplemented
-       , NotImplemented(..)
+       , Undefined (..)
        ) where
 
-import           Control.Monad          (Monad, return)
-import           Control.Monad.IO.Class (MonadIO)
-import           Data.Data              (Data)
-import           Data.Text              (Text, unpack)
-import           Data.Typeable          (Typeable)
-import           GHC.Generics           (Generic)
-import           System.IO.Unsafe       (unsafePerformIO)
+import           Control.Monad    (Monad, return)
+import           Data.Data        (Data)
+import           Data.Text        (Text, unpack)
+import           Data.Typeable    (Typeable)
+import           GHC.Generics     (Generic)
+import           System.IO.Unsafe (unsafePerformIO)
 
-import qualified Base                   as P
-import qualified Prelude                as P
-import           Show                   (Print, putStrLn)
+import qualified Base             as P
+import qualified Prelude          as P
+import           Show             (Print, putStrLn)
 
-import           Applicative            (pass)
+import           Applicative      (pass)
 
 {-# WARNING trace "'trace' remains in code" #-}
 trace :: Print b => b -> a -> a
@@ -37,13 +34,6 @@ trace string expr = unsafePerformIO (do
     putStrLn string
     return expr)
 
-{-# WARNING traceIO "'traceIO' remains in code" #-}
-traceIO :: (Print b, MonadIO m) => b -> a -> m a
-traceIO string expr = do
-    putStrLn string
-    return expr
-
-{-# WARNING error "'error' remains in code (or use 'panic')" #-}
 error :: Text -> a
 error s = P.error (unpack s)
 
@@ -67,14 +57,10 @@ traceM s = trace (unpack s) pass
 traceId :: Text -> Text
 traceId s = trace s s
 
-{-# WARNING notImplemented "'notImplemented' remains in code" #-}
-notImplemented :: a
-notImplemented = P.error "Not implemented"
+{-# WARNING Undefined "'Undefined' type remains in code" #-}
+data Undefined = Undefined
+    deriving (P.Eq, P.Ord, P.Show, P.Read, P.Enum, P.Bounded, Data, Typeable, Generic)
 
-{-# WARNING NotImplemented "'NotImplemented' remains in code" #-}
-data NotImplemented = NotImplemented
-    deriving (P.Eq, P.Ord, P.Show, Data, Typeable, Generic)
-
-{-# WARNING undefined "'undefined' remains in code (or use 'panic')" #-}
+{-# WARNING undefined "'undefined' function remains in code (or use 'error')" #-}
 undefined :: a
 undefined = P.undefined